+++ sample/config/main.php
@@ -77,6 +77,7 @@ return [
             'codesEncryptionKey' => getenv('YII2_OAUTH2_SERVER_CODES_ENCRYPTION_KEY'), // The encryption key for authorization and refresh codes.
             'storageEncryptionKeys' => getenv('YII2_OAUTH2_SERVER_STORAGE_ENCRYPTION_KEYS'), // The encryption key for storage like client secrets.
             'defaultStorageEncryptionKey' => '2022-01-01', // The index of the default key in storageEncryptionKeys.
+            'nonTlsAllowedRanges' => YII_DEBUG ? ['localhost', 'private'] : 'localhost', // By default, Clients are only allowed to connect using TLS, ranges specified here are exempted form this restriction.
             'grantTypes' => [ // For more information which grant types to use, please see https://oauth2.thephpleague.com/authorization-server/which-grant/.
                 Oauth2Module::GRANT_TYPE_AUTH_CODE,
                 Oauth2Module::GRANT_TYPE_CLIENT_CREDENTIALS,